refactor: ts of contract replay
