+++ packages/stores/src/mockStore.ts
@@ -319,6 +319,7 @@ const mock = (): TStores & { is_mock: boolean } => {
             is_cashier_visible: false,
             is_wallet_modal_visible: false,
             is_chart_layout_default: false,
+            is_chart_countdown_visible: false,
             is_closing_create_real_account_modal: false,
             is_dark_mode_on: false,
             is_language_settings_modal_on: false,
@@ -367,6 +368,7 @@ const mock = (): TStores & { is_mock: boolean } => {
             setAppContentsScrollRef: jest.fn(),
             shouldNavigateAfterChooseCrypto: jest.fn(),
             simple_duration_unit: 't',
+            toggleHistoryTab: jest.fn(),
             toggleLanguageSettingsModal: jest.fn(),
             togglePositionsDrawer: jest.fn(),
             toggleLinkExpiredModal: jest.fn(),
@@ -587,13 +589,31 @@ const mock = (): TStores & { is_mock: boolean } => {
         pushwoosh: {},
         contract_replay: {
             contract_store: {
+                accumulator_previous_spot_time: null,
+                barriers_array: [],
+                contract_config: {},
                 contract_info: {},
+                contract_update: {},
+                contract_update_history: [],
                 digits_info: {},
                 display_status: '',
+                getContractsArray: jest.fn(),
                 is_digit_contract: false,
                 is_ended: false,
+                marker: null,
+                markers_array: [],
             },
+            chart_state: '',
+            chartStateChange: jest.fn(),
             has_error: false,
+            is_chart_loading: true,
+            is_forward_starting: false,
+            is_market_closed: false,
+            is_sell_requested: false,
+            onClickCancel: jest.fn(),
+            onClickSell: jest.fn(),
+            onMount: jest.fn(),
+            onUnmount: jest.fn(),
             removeErrorMessage: jest.fn(),
             removeAccountSwitcherListener: jest.fn(),
             setAccountSwitcherListener: jest.fn(),
